+++ src/Commands/impl/PlayMemeCommand.ts
@@ -2,14 +2,18 @@ import { Message } from 'discord.js';
 import { Subscription } from '../../JazzyBot';
 import { CommandChain } from '../CommandChain';
 import { EmptyCommand } from './EmptyCommand';
+import memesPlaylist from '../../MusicPlayer/playlists/memes.json';
+import { randomIndex } from '../../utils';
 
 export class PlayMemeCommand implements CommandChain {
   private nextCommand: CommandChain;
   private commands: Set<string>;
+  private memes: string[];
 
   constructor() {
     this.nextCommand = new EmptyCommand();
     this.commands = new Set<string>(['meme']);
+    this.memes = memesPlaylist.map((meme) => meme.url);
   }
 
   public setNext(nextCommand: CommandChain) {
@@ -20,6 +24,7 @@ export class PlayMemeCommand implements CommandChain {
     if (!this.commands.has(command))
       return this.nextCommand.processCommand(command, message, subscription);
 
-    subscription.musicPlayer.playMeme();
+    const memeIndex = randomIndex(this.memes);
+    subscription.musicPlayer.play(this.memes[memeIndex]);
   }
 }
