+++ src/Commands/impl/PlayMemesCommand.ts
@@ -2,14 +2,18 @@ import { Message } from 'discord.js';
 import { Subscription } from '../../JazzyBot';
 import { CommandChain } from '../CommandChain';
 import { EmptyCommand } from './EmptyCommand';
+import memesPlaylist from '../../MusicPlayer/playlists/memes.json';
+import { shuffle } from '../../utils';
 
 export class PlayMemesCommand implements CommandChain {
   private nextCommand: CommandChain;
   private commands: Set<string>;
+  private memes: string[];
 
   constructor() {
     this.nextCommand = new EmptyCommand();
     this.commands = new Set<string>(['memes']);
+    this.memes = memesPlaylist.map((meme) => meme.url);
   }
 
   public setNext(nextCommand: CommandChain) {
@@ -20,6 +24,7 @@ export class PlayMemesCommand implements CommandChain {
     if (!this.commands.has(command))
       return this.nextCommand.processCommand(command, message, subscription);
 
-    await subscription.musicPlayer.playMemes();
+    this.memes = shuffle(this.memes);
+    subscription.musicPlayer.playPlaylist(this.memes);
   }
 }
