+++ src/Commands/impl/PlaySambaCommand.ts
@@ -2,14 +2,18 @@ import { Message } from 'discord.js';
 import { Subscription } from '../../JazzyBot';
 import { CommandChain } from '../CommandChain';
 import { EmptyCommand } from './EmptyCommand';
+import { parseVideoInfo, randomIndex } from '../../utils';
+import sambasPlaylist from '../../MusicPlayer/playlists/sambas.json';
 
 export class PlaySambaCommand implements CommandChain {
   private nextCommand: CommandChain;
   private commands: Set<string>;
+  private sambas: string[];
 
   constructor() {
     this.nextCommand = new EmptyCommand();
     this.commands = new Set<string>(['samba']);
+    this.sambas = sambasPlaylist.map((samba) => samba.url);
   }
 
   public setNext(nextCommand: CommandChain) {
@@ -20,6 +24,7 @@ export class PlaySambaCommand implements CommandChain {
     if (!this.commands.has(command))
       return this.nextCommand.processCommand(command, message, subscription);
 
-    subscription.musicPlayer.playSamba();
+    const sambaIndex = randomIndex(this.sambas);
+    subscription.musicPlayer.play(this.sambas[sambaIndex]);
   }
 }
