+++ src/Commands/impl/PlaySambasCommand.ts
@@ -2,14 +2,18 @@ import { Message } from 'discord.js';
 import { Subscription } from '../../JazzyBot';
 import { CommandChain } from '../CommandChain';
 import { EmptyCommand } from './EmptyCommand';
+import sambasPlaylist from '../../MusicPlayer/playlists/sambas.json';
+import { shuffle } from '../../utils';
 
 export class PlaySambasCommand implements CommandChain {
   private nextCommand: CommandChain;
   private commands: Set<string>;
+  private sambas: string[];
 
   constructor() {
     this.nextCommand = new EmptyCommand();
     this.commands = new Set<string>(['sambas']);
+    this.sambas = sambasPlaylist.map((samba) => samba.url);
   }
 
   public setNext(nextCommand: CommandChain) {
@@ -20,6 +24,7 @@ export class PlaySambasCommand implements CommandChain {
     if (!this.commands.has(command))
       return this.nextCommand.processCommand(command, message, subscription);
 
-    await subscription.musicPlayer.playSambaPlaylist();
+    this.sambas = shuffle(this.sambas);
+    subscription.musicPlayer.playPlaylist(this.sambas);
   }
 }
